Enum oxcable_subtractive_synth::Message

pub enum Message {
    SetGain(f32),
    SetOsc1(Waveform),
    SetOsc2(Waveform),
    SetOsc1Transpose(f32),
    SetOsc2Transpose(f32),
    SetAttack(f32),
    SetDecay(f32),
    SetSustain(f32),
    SetRelease(f32),
    SetLFOFreq(f32),
    SetVibrato(f32),
    SetTremolo(f32),
    SetFilterFirstOrder(FilterMode),
    SetFilterSecondOrder(FilterMode),
    SendMidiEvent(MidiEvent),
}

Defines the messages that the synthesizer supports.

Variants

SetGain

Sets the output gain, in decibels.

SetOsc1

Sets the waveform for the first oscillator.

SetOsc2

Sets the waveform for the second oscillator.

SetOsc1Transpose

Sets the transposition of the first oscillator, in steps.

SetOsc2Transpose

Sets the transposition of the second oscillator, in steps.

SetAttack

Sets the ADSR attack duration, in seconds.

SetDecay

Sets the ADSR decay duration, in seconds.

SetSustain

Sets the ADSR sustain level, from 0 to 1.

SetRelease

Sets the ADSR release duration, in seconds.

SetLFOFreq

Sets the LFO frequency, in Hz.

SetVibrato

Sets the vibrato intensity, in steps.

SetTremolo

Sets the tremolo intensity, in decibels.

SetFilterFirstOrder

Sets the filter to a first order filter of the specified mode.

SetFilterSecondOrder

Sets the filter to a second order filter of the specified mode.

SendMidiEvent

Sends the provided MIDI event to the synth.